[리팩토링] 주석 추가, 사용되지 않는 함수 제거
-  CoreData Driven Model Group 생성 (Without folder)
- 사용하지 않는 Tag 제거
- DateStringSet 주석 추가, 리팩토링
