+++ OneDay/OneDay/AppDelegate/AppDelegate.swift
@@ -20,24 +20,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         LocationService.service.requestAuth()
         return true
     }
-
-    func applicationWillResignActive(_ application: UIApplication) {
-        // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
-        // Use this method to pause ongoing tasks, disable timers, and invalidate graphics rendering callbacks. Games should use this method to pause the game.
-    }
-
+    
     func applicationDidEnterBackground(_ application: UIApplication) {
-    }
-
-    func applicationWillEnterForeground(_ application: UIApplication) {
-        // Called as part of the transition from the background to the active state; here you can undo many of the changes made on entering the background.
-    }
-
-    func applicationDidBecomeActive(_ application: UIApplication) {
-        // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
+        CoreDataManager.shared.save()
     }
 
     func applicationWillTerminate(_ application: UIApplication) {
+        CoreDataManager.shared.save()
     }
-
 }
