+++ OneDay/OneDay/Model/DataStringSet.swift
@@ -7,7 +7,10 @@
 //
 
 import Foundation
-    
+
+/**
+ Date에서 사용할 day, weekday 등을 저장하는 구조ㅊ
+ */
 struct DateStringSet {
     let weekDay: String
     let time: String
@@ -17,9 +20,8 @@ struct DateStringSet {
     init(date input: Date?) {
         let date: Date = input ?? Date()
         
-        let dateFormatter = DateFormatter()
+        let dateFormatter = DateFormatter.defaultInstance
         dateFormatter.locale = Locale.init(identifier: "ko")
-        dateFormatter.timeZone = TimeZone.current
         
         dateFormatter.dateStyle = .full
         dateFormatter.timeStyle = .none
