+++ OneDay/OneDay/Model/Entry+CoreDataClass.swift
@@ -12,4 +12,21 @@ import CoreData
 @objc(Entry)
 public class Entry: NSManagedObject {
     
+    func updateDate(date: Date) {
+        self.date = date
+        let dateComponents = Calendar.current.dateComponents(in: TimeZone.current, from: date)
+        if let month = dateComponents.month as NSNumber? {
+            self.month = month
+        }
+        if let day = dateComponents.day as NSNumber? {
+            self.day = day
+        }
+        if let year = dateComponents.year as NSNumber? {
+            self.year = year
+        }
+        self.monthAndYear = "\(year) \(month)"
+    }
+    
+    lazy var thmbnailFileName: String = "entry_thumbnail_\(entryId.uuidString)"
+    
 }
