--- OneDay/OneDay/Model/Tag+CoreDataClass.swift
@@ -1,16 +1,0 @@
-//
-//  Tag+CoreDataClass.swift
-//  OneDay
-//
-//  Created by juhee on 13/02/2019.
-//  Copyright © 2019 teamA2. All rights reserved.
-//
-//
-
-import Foundation
-import CoreData
-
-@objc(Tag)
-public class Tag: NSManagedObject {
-
-}
